+++ bin/stub-server.test.ts
@@ -12,6 +12,7 @@ const correctPort = '16928';
 const EXIT_SUCCESS = 0;
 const EXIT_FAILURE = 1;
 
+// eslint-disable-next-line unicorn/no-null
 const EXIT_SIGTERM = null;
 
 const killStubServerAfterRunning = (process: ChildProcessWithoutNullStreams) =>
@@ -20,9 +21,9 @@ const killStubServerAfterRunning = (process: ChildProcessWithoutNullStreams) =>
 // [Remove all ANSI colors/styles from strings](https://stackoverflow.com/q/25245716)
 const cleanAnsi = (str: string) =>
   // eslint-disable-next-line no-control-regex
-  str.replace(/[\u001b\u009b][[()#;?]*(?:[0-9]{1,4}(?:;[0-9]{0,4})*)?[0-9A-ORZcf-nqry=><]/g, '');
+  str.replace(/[\u001B\u009B][#();?[]*(?:\d{1,4}(?:;\d{0,4})*)?[\d<=>A-ORZcf-nqry]/g, '');
 
-// eslint-disable-next-line jest/no-test-callback
+// eslint-disable-next-line jest/no-done-callback
 test('correct config param', done => {
   expect.assertions(2);
 
@@ -43,7 +44,7 @@ test('correct config param', done => {
   killStubServerAfterRunning(process);
 });
 
-// eslint-disable-next-line jest/no-test-callback
+// eslint-disable-next-line jest/no-done-callback
 test('correct config and port params', done => {
   expect.assertions(2);
 
@@ -64,7 +65,7 @@ test('correct config and port params', done => {
   killStubServerAfterRunning(process);
 });
 
-// eslint-disable-next-line jest/no-test-callback
+// eslint-disable-next-line jest/no-done-callback
 test('network request', done => {
   // Unfortunately it does not test CORS because the request is performed server side :-/
   // Users perform requests in a web browser where CORS is enabled
@@ -81,8 +82,10 @@ test('network request', done => {
         });
 
         res.on('close', () => {
+          /* eslint-disable jest/no-conditional-expect */
           expect(res.statusCode).toEqual(200);
           expect(JSON.parse(resData)).toEqual({ stub: 'GET_200_OK.json' });
+          /* eslint-enable jest/no-conditional-expect */
 
           process.kill();
         });
@@ -98,7 +101,7 @@ test('network request', done => {
   });
 });
 
-// eslint-disable-next-line jest/no-test-callback
+// eslint-disable-next-line jest/no-done-callback
 test('incorrect config param', done => {
   expect.assertions(2);
 
@@ -115,7 +118,7 @@ test('incorrect config param', done => {
   });
 });
 
-// eslint-disable-next-line jest/no-test-callback
+// eslint-disable-next-line jest/no-done-callback
 test('incorrect port param', done => {
   expect.assertions(2);