+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+- Update npm packages
+- Use eslint-plugin-unicorn
+
 ## 0.3.2 (2020/08/11)
 
 - Update npm packages